Add sub-clerk feature
Looking at the memory problems, it is possible there are issues with the
garbage collector due to the clerk being a permanent process so it
doesn't dereference all the binary objects it passes through.

There was a similar problem with the Penciller's clerk which was
resolved by getting the destination file process to do all the binary
handling.

In this case, trying a simpler way of reolving this, by getting a
separate process to do the actual file compaction - one that always
closes on completion.