Set waste_fp
If a file is started as it was once the active journal, it will have
the waste_fp - and so will be moved to waste when it closes out of
delete_pending.  However, files started from the manifest on startup
(i.e. as they already existed before the store was started) were not
given a waste_fp and so were automatically deleted.