Memory management hypothesis
One of the issues in volume testing has been the relative importance of
the time spent copying the in-memory skiplists between processes,
compared to the lookup time for finding things within the lists.

What is we converted all the sub-elements of the skiplist into binaries?
All these cmponents would then be in the heap and referenced, rather
than in the process memory - so in theory copying between processes
should now be largely a transfer of references.

The lookup of an individual item would be slower, but the lookup is
already protected by a hash style filter ... so perhaps this doesn't
matter so much.