Changes to run in OTP18
This didn’t work for me.  I added a uniform_s function for the if not
old_rand case, as well as using strong_rand_bytes in old as well as new
(I couldn’t find crypto:strong_rand/1).  Was this meant to be plain
rand_bytes/1.

Might be being stupid, and missed something that Heinz is doing here.
With these changes rebar3 dialyzer/eunit/ct all work in OTP18 again.